Break up evp.h into smaller header files
